Refine view status formatting and reuse HTTP client
