+++ seva/viewmodels/runs_vm.py
@@ -5,6 +5,7 @@ from datetime import datetime
 from typing import Any, Dict, List, Optional
 
 from seva.domain.runs_registry import RunEntry, RunsRegistry
+from .status_format import registry_status_label
 
 
 @dataclass
@@ -59,15 +60,7 @@ class RunsVM:
         )
 
     def _format_status(self, entry: RunEntry) -> str:
-        if entry.status == "done":
-            return "Done (Downloaded)" if entry.download.done else "Done"
-        if entry.status == "cancelled":
-            return "Cancelled"
-        if entry.status == "error":
-            return "Error"
-        if entry.status == "deleted":
-            return "Deleted"
-        return "Running"
+        return registry_status_label(entry.status, downloaded=entry.download.done)
 
     def _format_progress(self, snapshot: Optional[Dict[str, Any]]) -> str:
         if not snapshot:
